+++ CHANGELOG.md
@@ -70,3 +70,9 @@ GET    /products?page=1&limit=10&title=jornada      -> Get all products with pag
 ## 1.0.16
 
 - Fix port that is presented at server start.
+
+
+## 1.1.0
+
+- Token access time fix
+- Add refresh token support for login system